Extract ClientExtensions to its own module
